:shirt: Improve source codes by using TypeScript
